+++ Isolator.Fody/ModuleWeaver.cs
@@ -23,10 +23,10 @@ public sealed partial class ModuleWeaver : BaseModuleWeaver
         WriteInfo($"{GetType().Assembly.GetName().Name} v{GetType().Assembly.GetVersion()}");
 
         FindMsCoreReferences();
-        ImportAssemblyLoader();
+        ImportAssemblyLoader(config.EnableDebug);
 
         FindContextNameMethod(config.ContextName);
-        FindLogMethod(config.EnableDebug);
+        //FindLogMethod(config.EnableDebug);
         CallAttach(config);
 
         IsolatorExecute();