+++ CHANGELOG.md
@@ -36,6 +36,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `IsolatorAvailable` to run `IsolatorExecuteRemoveAttributes` to remove attributes when available.
 - Update `Configuration` with `SkipIsolator` to skip isolator for specific assemblies.
 - Update `SkipIsolator` to show warning log when assembly is skipped.
+- Create `Isolator.Template.Debug` project to hold debug with `DEBUG` level logs.
+- Update `AssemblyLoaderImporter` to select `debug` or `release` version of `Isolator.Template` based on `EnableDebug` setting.
 
 [vNext]: ../../compare/1.0.0...HEAD
 [1.0.0]: ../../compare/1.0.0